requirements.txt
matplotlib==3.3.3
pandas==1.1.5
seaborn==0.11.0
torchvision==0.8.2
requests==2.26.0
torch==1.7.1
decord==0.5.2
tqdm==4.56.0
tensorflow_gpu==2.3.1
opencv_python_headless==4.4.0.46
numpy==1.23.4
Pillow==9.2.0
scikit_learn==1.1.2
tensorflow==2.10.0
